requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
PyPDF2
chromadb
faiss-cpu
langchain_google_genai